requirements.txt
absl-py==0.9.0
astor==0.8.1
cycler==0.10.0
deeplift==0.6.8.1
gast==0.3.3
grpcio==1.27.2
h5py==2.8.0
joblib==0.14.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
kneed==0.5.0
Markdown==3.2.1
matplotlib==3.0.3
numpy==1.16.0
powerlaw==1.4.4
protobuf==3.11.3
pyparsing==2.4.6
python-dateutil==2.8.1
PyYAML==5.3
scikit-learn==0.22.1
scipy==1.2.0
six==1.14.0
sklearn==0.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
Werkzeug==1.0.0
